requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
requests
loguru
langchain
langchain-community
langchain-openai
chromadb
PyYAML
browser-cookie3
PyJWT